requirements.txt
Flask==3.0.0
gTTS==2.3.2
piper-tts>=0.0.19
numpy>=1.24.0
soundfile>=0.12.0
scipy>=1.10.0
pygame>=2.5.0
google-genai==0.3.0
google-generativeai==0.3.0
requests==2.31.0
SQLAlchemy
pyttsx3